track nixpkgs addition of 0014 patch
To work around some compilation issues on an llvm bump up in
nixpkgs, someone added this patch for oil's py-yajl. We do
not actually *need* this patch in this repo, but I want to
avoid the confusion re-using this patch number could cause,
so I'm going to add it here to document its use, immediately
remove it in the next commit, and then use 0015 for the next
patch.